+++ src/mlpc/loss.c
@@ -2,11 +2,11 @@
 #include "loss.h"
 #include "matrix.h"
 
-double errorFunction(Matrix yhat, Matrix y, Matrix error)
+float errorFunction(Matrix yhat, Matrix y, Matrix error)
 {
     matrix_copy(error, y);
 
-    double sum = 0;
+    float sum = 0;
     int n = error.rows * error.columns;
     for (int i = 0; i < n; i++)
         sum += error.data[i];
@@ -14,16 +14,16 @@ double errorFunction(Matrix yhat, Matrix y, Matrix error)
     return sum / n;
 }
 
-double mse(Matrix yhat, Matrix y, Matrix error)
+float mse(Matrix yhat, Matrix y, Matrix error)
 {
     matrix_difference(yhat, y, error);
 
-    double sum = 0;
+    float sum = 0;
     int n = error.rows * error.columns;
     for (int i = 0; i < n; i++)
         sum += pow(error.data[i], 2);
     
-    return sum / (double)n;
+    return sum / (float)n;
 }
 
 LossFunction getLossFunction(int code)
